+++ src/App.tsx
@@ -4,16 +4,13 @@ import React, { useEffect, useRef, useState } from 'react';
 import { UserWarning } from './UserWarning';
 import { getTodos, USER_ID } from './api/todos';
 import { Todo } from './types/Todo';
-import cls from 'classnames';
+import { TodoHeader } from './component/TodoHeader/TodoHeader';
+import { TodoFooter } from './component/TodoFooter/TodoFooter';
+import { Status } from './types/Status';
+import { TodoList } from './component/TodoList/TodoList';
+import { TodoErrorNotification } from './component/TodoErrorNotification/TodoErrorNotification';
 
 export const App: React.FC = () => {
-
-
-  enum Status {
-    'all' = 'all',
-    'active' = 'active',
-    'completed' = 'completed',
-  }
 
   enum Errors {
     'titleError' = 'Title should not be empty',
@@ -22,22 +19,30 @@ export const App: React.FC = () => {
     'loadTodoError' = 'Unable to load todos',
   }
 
-  const [todos, setTodos] = useState<Todo[]>([]);
-  const [title, setTitle] = useState('');
+
   const [hasTitleError, setHasTitleError] = useState(false);
   const [loadTodoError, setLoadTodoError] = useState(false);
   const [status, setStatus] = useState('all');
   const errorTimerRef = useRef<number | null>(null);
+  const [todos, setTodos] = useState<Todo[]>([]);
 
   if (!USER_ID) {
     return <UserWarning />;
   }
 
+  useEffect(() => {
+    getTodos()
+      .then(setTodos)
+      .catch(() => {
+        setLoadTodoError(true);
+        resetError();
+      });
+  }, []);
+
   const resetError = () => {
     if (errorTimerRef.current) {
       clearTimeout(errorTimerRef.current);
     }
-
     errorTimerRef.current = window.setTimeout(() => {
       setHasTitleError(false);
       setLoadTodoError(false);
@@ -45,9 +50,6 @@ export const App: React.FC = () => {
     }, 3000);
   };
 
-  const reset = () => {
-    setTitle('');
-  };
 
   const hideErrors = () => {
     if (errorTimerRef.current) {
@@ -59,27 +61,14 @@ export const App: React.FC = () => {
     setLoadTodoError(false);
   };
 
-  useEffect(() => {
-    getTodos()
-      .then(setTodos)
-      .catch(() => {
-        setLoadTodoError(true);
-        resetError();
-      });
-  }, []);
-
-  const handleTitleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setTitle(event.target.value);
+  const getErrorMessage = () => {
+    if (hasTitleError) return Errors.titleError;
+    if (loadTodoError) return Errors.loadTodoError;
+    return '';
   };
+  const isErrorHidden = () => !hasTitleError && !loadTodoError;
 
-  const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-    if (!title) {
-      setHasTitleError(true);
-      resetError();
-      return;
-    }
-  };
+
 
   const filteredTodos = todos.filter(todo => {
     if (status === Status.active) {
@@ -95,151 +84,20 @@ export const App: React.FC = () => {
     <div className="todoapp">
       <h1 className="todoapp__title">todos</h1>
       <div className="todoapp__content">
-        <header className="todoapp__header">
-          {/* this button should have `active` class only if all todos are completed */}
-          <button
-            type="button"
-            className="todoapp__toggle-all active"
-            data-cy="ToggleAllButton"
-          />
-
-          {/* Add a todo on form submit */}
-          <form onSubmit={handleSubmit} onReset={reset}>
-            <input
-              data-cy="NewTodoField"
-              type="text"
-              value={title}
-              className="todoapp__new-todo"
-              placeholder="What needs to be done?"
-              onChange={handleTitleChange}
-            />
-          </form>
-        </header>
-
-        {filteredTodos.length > 0 && (
-          <section className="todoapp__main" data-cy="TodoList">
-            {/*This is a completed todo */}
-            {filteredTodos.map(todo => {
-              return (
-                <div
-                  data-cy="Todo"
-                  className={cls('todo', { completed: todo.completed })}
-                >
-                  <label className="todo__status-label">
-                    <input
-                      data-cy="TodoStatus"
-                      type="checkbox"
-                      className="todo__status"
-                      checked={todo.completed}
-                    />
-                  </label>
-
-                  <span data-cy="TodoTitle" className="todo__title">
-                    {todo.title}
-                  </span>
-
-                  {/* Remove button appears only on hover */}
-                  <button
-                    type="button"
-                    className="todo__remove"
-                    data-cy="TodoDelete"
-                  >
-                    ×
-                  </button>
-
-                  {/* overlay will cover the todo while it is being deleted or updated */}
-                  <div data-cy="TodoLoader" className="modal overlay">
-                    <div className="modal-background has-background-white-ter" />
-                    <div className="loader" />
-                  </div>
-                </div>
-              );
-            })}
-          </section>
-        )}
-
-        {/* Hide the footer if there are no todos */}
-
+        <TodoHeader setHasTitleError={setHasTitleError} />
         {todos.length > 0 && (
-          <footer className="todoapp__footer" data-cy="Footer">
-            <span className="todo-count" data-cy="TodosCounter">
-              {todos.filter(todo => !todo.completed).length} items left
-            </span>
-
-            {/* Active link should have the 'selected' class */}
-            <nav className="filter" data-cy="Filter">
-              <a
-                href="#/"
-                className={cls('filter__link', {
-                  selected: status === Status.all,
-                })}
-                data-cy="FilterLinkAll"
-                onClick={() => {
-                  setStatus('all');
-                }}
-              >
-                All
-              </a>
-
-              <a
-                href="#/active"
-                className={cls('filter__link', {
-                  selected: status === Status.active,
-                })}
-                data-cy="FilterLinkActive"
-                onClick={() => {
-                  setStatus(Status.active);
-                }}
-              >
-                Active
-              </a>
-
-              <a
-                href="#/completed"
-                className={cls('filter__link', {
-                  selected: status === Status.completed,
-                })}
-                data-cy="FilterLinkCompleted"
-                onClick={() => {
-                  setStatus(Status.completed);
-                }}
-              >
-                Completed
-              </a>
-            </nav>
-            {/* this button should be disabled if there are no completed todos */}
-            <button
-              type="button"
-              className="todoapp__clear-completed"
-              data-cy="ClearCompletedButton"
-            >
-              Clear completed
-            </button>
-          </footer>
+          <>
+            <TodoList filteredTodos={filteredTodos} />
+            <TodoFooter todos={todos} status={status} setStatus={setStatus} />
+          </>
         )}
+        {/* Hide the footer if there are no todos */}
       </div>
 
       {/* DON'T use conditional rendering to hide the notification */}
       {/* Add the 'hidden' class to hide the message smoothly */}
+      <TodoErrorNotification getErrorMessage={getErrorMessage} isErrorHidden={isErrorHidden} hideErrors={hideErrors} />
 
-      <div
-        data-cy="ErrorNotification"
-        className={cls(
-          'notification is-danger is-light has-text-weight-normal',
-          { hidden: !hasTitleError && !loadTodoError },
-        )}
-      >
-        <button
-          data-cy="HideErrorButton"
-          type="button"
-          className="delete"
-          onClick={hideErrors}
-        />
-        {/*/!* show only one message at a time *!/*/}
-
-        {hasTitleError && <p>{Errors.titleError}</p>}
-        {loadTodoError && <p>{Errors.loadTodoError}</p>}
-      </div>
     </div>
   );
 };
