+++ src/component/TodoHeader/TodoHeader.tsx
@@ -1,0 +1,47 @@
+import React, { useState } from 'react';
+
+
+export interface TodoHeaderProps {
+  setHasTitleError: (titleError: boolean) => void;
+}
+
+
+export const TodoHeader: React.FC<TodoHeaderProps>= ({setHasTitleError}) =>{
+  const [title ,setTitle ] = useState('')
+
+  const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
+    event.preventDefault();
+    if (!title) {
+      setHasTitleError(true);
+      return;
+    }
+  };
+
+  const handleTitleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    setTitle(event.target.value);
+  };
+
+  return (
+    <header className="todoapp__header">
+      {/* this button should have `active` class only if all todos are completed */}
+      <button
+        type="button"
+        className="todoapp__toggle-all active"
+        data-cy="ToggleAllButton"
+      />
+
+      {/* Add a todo on form submit */}
+      <form onSubmit={handleSubmit}
+            onReset={() => setTitle('')}>
+        <input
+          data-cy="NewTodoField"
+          type="text"
+          value={title}
+          className="todoapp__new-todo"
+          placeholder="What needs to be done?"
+          onChange={handleTitleChange}
+        />
+      </form>
+    </header>
+  )
+}
